move test setup to testutil
